Huge Resharper clean up.  Fixed up test project.json

+++ src/SharpCompress/Compressor/PPMd/I1/MemoryNode.cs
@@ -1,6 +1,6 @@
 #region Using
 
-using System;
+
 
 #endregion
 
@@ -50,15 +50,15 @@ namespace SharpCompress.Compressor.PPMd.I1
         {
             get
             {
-                return ((uint) Memory[Address]) | ((uint) Memory[Address + 1]) << 8 | ((uint) Memory[Address + 2]) << 16 |
-                       ((uint) Memory[Address + 3]) << 24;
+                return Memory[Address] | ((uint)Memory[Address + 1]) << 8 | ((uint)Memory[Address + 2]) << 16 |
+                       ((uint)Memory[Address + 3]) << 24;
             }
             set
             {
-                Memory[Address] = (byte) value;
-                Memory[Address + 1] = (byte) (value >> 8);
-                Memory[Address + 2] = (byte) (value >> 16);
-                Memory[Address + 3] = (byte) (value >> 24);
+                Memory[Address] = (byte)value;
+                Memory[Address + 1] = (byte)(value >> 8);
+                Memory[Address + 2] = (byte)(value >> 16);
+                Memory[Address + 3] = (byte)(value >> 24);
             }
         }
 
@@ -71,15 +71,15 @@ namespace SharpCompress.Compressor.PPMd.I1
             {
                 return
                     new MemoryNode(
-                        ((uint) Memory[Address + 4]) | ((uint) Memory[Address + 5]) << 8 |
-                        ((uint) Memory[Address + 6]) << 16 | ((uint) Memory[Address + 7]) << 24, Memory);
+                                   Memory[Address + 4] | ((uint)Memory[Address + 5]) << 8 |
+                                   ((uint)Memory[Address + 6]) << 16 | ((uint)Memory[Address + 7]) << 24, Memory);
             }
             set
             {
-                Memory[Address + 4] = (byte) value.Address;
-                Memory[Address + 5] = (byte) (value.Address >> 8);
-                Memory[Address + 6] = (byte) (value.Address >> 16);
-                Memory[Address + 7] = (byte) (value.Address >> 24);
+                Memory[Address + 4] = (byte)value.Address;
+                Memory[Address + 5] = (byte)(value.Address >> 8);
+                Memory[Address + 6] = (byte)(value.Address >> 16);
+                Memory[Address + 7] = (byte)(value.Address >> 24);
             }
         }
 
@@ -90,25 +90,22 @@ namespace SharpCompress.Compressor.PPMd.I1
         {
             get
             {
-                return ((uint) Memory[Address + 8]) | ((uint) Memory[Address + 9]) << 8 |
-                       ((uint) Memory[Address + 10]) << 16 | ((uint) Memory[Address + 11]) << 24;
+                return Memory[Address + 8] | ((uint)Memory[Address + 9]) << 8 |
+                       ((uint)Memory[Address + 10]) << 16 | ((uint)Memory[Address + 11]) << 24;
             }
             set
             {
-                Memory[Address + 8] = (byte) value;
-                Memory[Address + 9] = (byte) (value >> 8);
-                Memory[Address + 10] = (byte) (value >> 16);
-                Memory[Address + 11] = (byte) (value >> 24);
+                Memory[Address + 8] = (byte)value;
+                Memory[Address + 9] = (byte)(value >> 8);
+                Memory[Address + 10] = (byte)(value >> 16);
+                Memory[Address + 11] = (byte)(value >> 24);
             }
         }
 
         /// <summary>
         /// Gets whether there is a next memory node available.
         /// </summary>
-        public bool Available
-        {
-            get { return Next.Address != 0; }
-        }
+        public bool Available { get { return Next.Address != 0; } }
 
         /// <summary>
         /// Link in the provided memory node.
@@ -171,7 +168,7 @@ namespace SharpCompress.Compressor.PPMd.I1
         /// <returns></returns>
         public static MemoryNode operator +(MemoryNode memoryNode, int offset)
         {
-            memoryNode.Address = (uint) (memoryNode.Address + offset*Size);
+            memoryNode.Address = (uint)(memoryNode.Address + offset * Size);
             return memoryNode;
         }
 
@@ -183,7 +180,7 @@ namespace SharpCompress.Compressor.PPMd.I1
         /// <returns></returns>
         public static MemoryNode operator +(MemoryNode memoryNode, uint offset)
         {
-            memoryNode.Address += offset*Size;
+            memoryNode.Address += offset * Size;
             return memoryNode;
         }
 
@@ -195,7 +192,7 @@ namespace SharpCompress.Compressor.PPMd.I1
         /// <returns></returns>
         public static MemoryNode operator -(MemoryNode memoryNode, int offset)
         {
-            memoryNode.Address = (uint) (memoryNode.Address - offset*Size);
+            memoryNode.Address = (uint)(memoryNode.Address - offset * Size);
             return memoryNode;
         }
 
@@ -207,7 +204,7 @@ namespace SharpCompress.Compressor.PPMd.I1
         /// <returns></returns>
         public static MemoryNode operator -(MemoryNode memoryNode, uint offset)
         {
-            memoryNode.Address -= offset*Size;
+            memoryNode.Address -= offset * Size;
             return memoryNode;
         }
 
@@ -242,7 +239,7 @@ namespace SharpCompress.Compressor.PPMd.I1
         {
             if (obj is MemoryNode)
             {
-                MemoryNode memoryNode = (MemoryNode) obj;
+                MemoryNode memoryNode = (MemoryNode)obj;
                 return memoryNode.Address == Address;
             }
             return base.Equals(obj);
