+++ src/SharpCompress/Compressor/PPMd/I1/PpmState.cs
@@ -1,6 +1,6 @@
 #region Using
 
-using System;
+
 
 #endregion
 
@@ -38,20 +38,12 @@ namespace SharpCompress.Compressor.PPMd.I1
         /// <summary>
         /// Gets or sets the symbol.
         /// </summary>
-        public byte Symbol
-        {
-            get { return Memory[Address]; }
-            set { Memory[Address] = value; }
-        }
+        public byte Symbol { get { return Memory[Address]; } set { Memory[Address] = value; } }
 
         /// <summary>
         /// Gets or sets the frequency.
         /// </summary>
-        public byte Frequency
-        {
-            get { return Memory[Address + 1]; }
-            set { Memory[Address + 1] = value; }
-        }
+        public byte Frequency { get { return Memory[Address + 1]; } set { Memory[Address + 1] = value; } }
 
         /// <summary>
         /// Gets or sets the successor.
@@ -62,15 +54,15 @@ namespace SharpCompress.Compressor.PPMd.I1
             {
                 return
                     new Model.PpmContext(
-                        ((uint) Memory[Address + 2]) | ((uint) Memory[Address + 3]) << 8 |
-                        ((uint) Memory[Address + 4]) << 16 | ((uint) Memory[Address + 5]) << 24, Memory);
+                                         Memory[Address + 2] | ((uint)Memory[Address + 3]) << 8 |
+                                         ((uint)Memory[Address + 4]) << 16 | ((uint)Memory[Address + 5]) << 24, Memory);
             }
             set
             {
-                Memory[Address + 2] = (byte) value.Address;
-                Memory[Address + 3] = (byte) (value.Address >> 8);
-                Memory[Address + 4] = (byte) (value.Address >> 16);
-                Memory[Address + 5] = (byte) (value.Address >> 24);
+                Memory[Address + 2] = (byte)value.Address;
+                Memory[Address + 3] = (byte)(value.Address >> 8);
+                Memory[Address + 4] = (byte)(value.Address >> 16);
+                Memory[Address + 5] = (byte)(value.Address >> 24);
             }
         }
 
@@ -80,10 +72,7 @@ namespace SharpCompress.Compressor.PPMd.I1
         /// </summary>
         /// <param name="offset"></param>
         /// <returns></returns>
-        public PpmState this[int offset]
-        {
-            get { return new PpmState((uint) (Address + offset*Size), Memory); }
-        }
+        public PpmState this[int offset] { get { return new PpmState((uint)(Address + offset * Size), Memory); } }
 
         /// <summary>
         /// Allow a pointer to be implicitly converted to a PPM state.
@@ -103,7 +92,7 @@ namespace SharpCompress.Compressor.PPMd.I1
         /// <returns></returns>
         public static PpmState operator +(PpmState state, int offset)
         {
-            state.Address = (uint) (state.Address + offset*Size);
+            state.Address = (uint)(state.Address + offset * Size);
             return state;
         }
 
@@ -126,7 +115,7 @@ namespace SharpCompress.Compressor.PPMd.I1
         /// <returns></returns>
         public static PpmState operator -(PpmState state, int offset)
         {
-            state.Address = (uint) (state.Address - offset*Size);
+            state.Address = (uint)(state.Address - offset * Size);
             return state;
         }
 
@@ -194,7 +183,7 @@ namespace SharpCompress.Compressor.PPMd.I1
         {
             if (obj is PpmState)
             {
-                PpmState state = (PpmState) obj;
+                PpmState state = (PpmState)obj;
                 return state.Address == Address;
             }
             return base.Equals(obj);
